requirements.txt
pyyaml==6.0.1
pillow==9.5.0
func_timeout==4.3.5
opencv-python==4.8.1.78
numpy==1.24.4
easyocr==1.7.1
